feat: add MicroPython supports
this feature/fix allows to run the library on MicroPython, which is
usefull when using embedded devices.

The MicroPython library does not support the `debug` call fully,
therefore it is removed by a pseudo function not printing anything. This
seems reasonable as most embedded devices are not constantly connected
to a serial port.

The MicroPython function `filter` does not support the `default` param,
therefore it catches the `StopIteration` and then returns `None`.

Also added a description in the `README.md` on the installation of the
missing `base64` library.

Signed-off-by: Paul Spooren <mail@aparcar.org>